yo dawg we put fix in yo fix

+++ LibraryDX12/Source/SceneImpl.cpp
@@ -55,7 +55,7 @@ void SceneImpl::render(ApplicationImpl &application, SwapChain &swapChain) {
     commandList.setGraphicsRootSignature(application.getPipelineStateController().getRootSignature(PipelineStateController::Identifier::PIPELINE_STATE_DEFAULT));
 
     // View projection matrix
-    float aspectRatio = swapChain.getWidth() / swapChain.getHeight();
+    float aspectRatio = (float) swapChain.getWidth() / swapChain.getHeight();
     camera->setAspectRatio(aspectRatio);
     const XMMATRIX viewMatrix = camera->getViewMatrix();
     const XMMATRIX projectionMatrix = camera->getProjectionMatrix();
